Update default version to 1.5.0 and URL in installation scripts
This commit modifies the .scripts/install-graal-ubuntu.sh, .scripts/install.sh, and .scripts/install-graal-macos.sh files. It updates the default app version from 1.4.0 to 1.5.0 and corrects the case of the username in the download URL. This is done to reflect the release of maven2sbt-cli version 1.5.0 and handle a possible case sensitivity issue that could prevent successful downloading of the package.
[skip ci]